+++ Tests/Primes_IsPrimeShould.cs
@@ -9,7 +9,7 @@ namespace Tests
         [Theory]
         [InlineData(0)]
         [InlineData(1)]
-        public void ReturnFalseGivenValuesLessThan2(uint number)
+        public void ReturnFalseGivenValuesLessThan2(ulong number)
         {
             bool result = Primes.IsPrime(number);
             Assert.False(result, number.ToString() + " should not be prime");
@@ -24,7 +24,7 @@ namespace Tests
         [InlineData(103)]
         [InlineData(181)]
         [InlineData(67)]
-        public void RecognizePrimeNumbers(uint number)
+        public void RecognizePrimeNumbers(ulong number)
         {
             bool result = Primes.IsPrime(number);
             Assert.True(result, number.ToString() + " should be prime");
@@ -38,7 +38,7 @@ namespace Tests
         [InlineData(100)]
         [InlineData(200)]
         [InlineData(148)]
-        public void RecognizeCompositeNumbers(uint number)
+        public void RecognizeCompositeNumbers(ulong number)
         {
             bool result = Primes.IsPrime(number);
             Assert.False(result, number.ToString() + "should not be prime");
